+++ lib/providers.dart
@@ -2,7 +2,6 @@ import 'package:flutter_dictionary/controllers/custom_drop_down_controller.dart'
 import 'package:flutter_dictionary/controllers/fetch_words_repository_controller.dart';
 import 'package:flutter_dictionary/controllers/package_info_repository_controller.dart';
 import 'package:flutter_dictionary/controllers/primary_view_controller.dart';
-import 'package:flutter_dictionary/controllers/tab_controller.dart';
 import 'package:flutter_dictionary/controllers/random_words_repository_controller.dart';
 import 'package:flutter_dictionary/controllers/words_repository_controller.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
