+++ lib/themeprovider.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-// import 'package:riverpod/riverpod.dart';
-
-// final themeprovider = ChangeNotifierProvider((ref) => ThemeSetter());
+import 'package:fluttertoast/fluttertoast.dart';
 
 final themeProvider = ChangeNotifierProvider((ref) => themeNotifier());
 
@@ -12,30 +10,37 @@ class themeNotifier extends ChangeNotifier {
 
   void setDark() {
     _value = ThemeMode.dark;
+    Fluttertoast.showToast(
+        msg: "Theme Set to Dark Mode",
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
+        timeInSecForIosWeb: 1,
+        textColor: Colors.white,
+        fontSize: 16.0);
     notifyListeners();
   }
+
   void setLight() {
     _value = ThemeMode.light;
+    Fluttertoast.showToast(
+        msg: "Theme Set to Light Mode",
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
+        timeInSecForIosWeb: 1,
+        textColor: Colors.white,
+        fontSize: 16.0);
     notifyListeners();
   }
+
   void setSystem() {
     _value = ThemeMode.system;
+    Fluttertoast.showToast(
+        msg: "Theme Set to System Mode",
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
+        timeInSecForIosWeb: 1,
+        textColor: Colors.white,
+        fontSize: 16.0);
     notifyListeners();
   }
 }
-
-
-
-// class ThemeSetter extends ChangeNotifier {
-//   ThemeMode _value = ThemeMode.light;
-//   ThemeMode get currentTheme => _value;
-
-//   void setDark() {
-//     _value = ThemeMode.dark;
-//     notifyListeners();
-//   }
-//   void setLight() {
-//     _value = ThemeMode.light;
-//     notifyListeners();
-//   }
-// }
