+++ lib/ui/components/home/about_view.dart
@@ -6,7 +6,6 @@ import 'package:flutter_dictionary/widgets/cached_net_image.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:flutter_dictionary/themeprovider.dart';
 
 class AboutView extends StatefulWidget {
   @override
@@ -18,18 +17,6 @@ class _AboutViewState extends State<AboutView> {
   @override
   Widget build(BuildContext context) {
     return Consumer(builder: (context, ref, _) {
-      void _setDark(BuildContext context) {
-        ref.read(themeProvider).setDark();
-      }
-
-      void _setLight(BuildContext context) {
-        ref.read(themeProvider).setLight();
-      }
-
-       void _setSystem(BuildContext context) {
-        ref.read(themeProvider).setSystem();
-      }
-
       return Padding(
         padding: const EdgeInsets.all(10.0),
         child: Column(
@@ -141,68 +128,6 @@ class _AboutViewState extends State<AboutView> {
                   ),
                 ],
               ),
-            ),
-            Container(
-              alignment: Alignment.centerLeft,
-              padding: const EdgeInsets.all(8.0),
-              child: Text(
-                "Settings",
-                style: GoogleFonts.poppins(
-                  fontSize: 30.0,
-                  fontWeight: FontWeight.w500,
-                ),
-              ),
-            ),
-            Row(
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(right: 8.0, left: 8.0),
-                  child: Text(
-                    "Theme:",
-                    style: GoogleFonts.poppins(
-                      fontSize: 20.0,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                ),
-                DropdownButton<String>(
-                  value: dropdownValue,
-                  items: <String>['System', 'Light', 'Dark']
-                      .map<DropdownMenuItem<String>>((String value) {
-                    return DropdownMenuItem<String>(
-                      value: value,
-                      child: Text(
-                        value,
-                        style: GoogleFonts.poppins(
-                          fontSize: 20.0,
-                          fontWeight: FontWeight.w500,
-                        ),
-                      ),
-                    );
-                  }).toList(),
-                  onChanged: (String? newValue) {
-                    if (newValue == 'System') {
-                      setState(() {
-                        dropdownValue = newValue!;
-                        _setSystem(context);
-
-                      });
-                    }
-                    if (newValue == 'Light') {
-                      setState(() {
-                        dropdownValue = newValue!;
-                        _setLight(context);
-                      });
-                    }
-                    if (newValue == 'Dark') {
-                      setState(() {
-                        dropdownValue = newValue!;
-                        _setDark(context);
-                      });
-                    }
-                  },
-                ),
-              ],
             ),
           ],
         ),
