+++ pubspec.yaml
@@ -23,6 +23,8 @@ dependencies:
   audioplayers: ^5.0.0
   uuid: ^3.0.7
   package_info_plus: ^4.1.0
+  fluttertoast: ^8.1.1
+
 
 dev_dependencies:
   flutter_test:
